+++ bindings/python/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = 'miraitensors'
-requires-python = '>=3.7'
+requires-python = '>=3.9'
 authors = [
     { name = 'Nicolas Patry', email = 'patry.nicolas@protonmail.com' },
     { name = 'Vladimir Vlasiuk', email = 'matterai.net@gmail.com' },
@@ -30,13 +30,13 @@ Source = 'https://github.com/huggingface/safetensors'
 
 [project.optional-dependencies]
 numpy = ["numpy>=1.21.6"]
-torch = ["safetensors[numpy]", "torch>=1.10"]
-tensorflow = ["safetensors[numpy]", "tensorflow>=2.11.0"]
+torch = ["miraitensors[numpy]", "torch>=1.10"]
+tensorflow = ["miraitensors[numpy]", "tensorflow>=2.11.0"]
 # pinning tf version 2.11.0 for doc-builder
-pinned-tf = ["safetensors[numpy]", "tensorflow==2.18.0"]
-jax = ["safetensors[numpy]", "flax>=0.6.3", "jax>=0.3.25", "jaxlib>=0.3.25"]
+pinned-tf = ["miraitensors[numpy]", "tensorflow==2.18.0"]
+jax = ["miraitensors[numpy]", "flax>=0.6.3", "jax>=0.3.25", "jaxlib>=0.3.25"]
 mlx = ["mlx>=0.0.9"]
-paddlepaddle = ["safetensors[numpy]", "paddlepaddle>=2.4.1"]
+paddlepaddle = ["miraitensors[numpy]", "paddlepaddle>=2.4.1"]
 quality = [
     "black==22.3",   # after updating to black 2023, also update Python version in pyproject.toml to 3.7
     "click==8.0.4",
@@ -44,7 +44,7 @@ quality = [
     "flake8>=3.8.3",
 ]
 testing = [
-    "safetensors[numpy]",
+    "miraitensors[numpy]",
     "h5py>=3.7.0",
     "huggingface_hub>=0.12.1",
     "setuptools_rust>=1.5.2",
@@ -54,15 +54,15 @@ testing = [
     "hypothesis>=6.70.2",
 ]
 all = [
-    "safetensors[torch]",
-    "safetensors[numpy]",
-    "safetensors[pinned-tf]",
-    "safetensors[jax]",
-    "safetensors[paddlepaddle]",
-    "safetensors[quality]",
-    "safetensors[testing]",
+    "miraitensors[torch]",
+    "miraitensors[numpy]",
+    "miraitensors[pinned-tf]",
+    "miraitensors[jax]",
+    "miraitensors[paddlepaddle]",
+    "miraitensors[quality]",
+    "miraitensors[testing]",
 ]
-dev = ["safetensors[all]"]
+dev = ["miraitensors[all]"]
 
 [tool.twine.repository]
 url = "https://europe-west2-python.pkg.dev/mirai-430618/pypi-pypi/"
@@ -74,7 +74,7 @@ build-backend = "maturin"
 
 [tool.maturin]
 python-source = "py_src"
-module-name = "safetensors._safetensors_rust"
+module-name = "miraitensors._safetensors_rust"
 bindings = 'pyo3'
 features = ["pyo3/extension-module"]
