+++ bindings/python/tests/test_simple.py
@@ -8,11 +8,11 @@ import jax.numpy as jnp
 import numpy as np
 import torch
 
-from safetensors import SafetensorError, safe_open, serialize
-from safetensors.numpy import load, load_file, save, save_file
-from safetensors.torch import _find_shared_tensors, storage_ptr, storage_size
-from safetensors.torch import load_file as load_file_pt
-from safetensors.torch import save_file as save_file_pt
+from miraitensors import SafetensorError, safe_open, serialize
+from miraitensors.numpy import load, load_file, save, save_file
+from miraitensors.torch import _find_shared_tensors, storage_ptr, storage_size
+from miraitensors.torch import load_file as load_file_pt
+from miraitensors.torch import save_file as save_file_pt
 
 
 class TestCase(unittest.TestCase):
